+++ ToSIC_SexyContent/ToSic.Sxc/Data/IDynamicEntity.cs
@@ -13,7 +13,7 @@ namespace ToSic.Sxc.Data
     /// The underlying IEntity <see cref="IEntity"/> is in the Entity property. 
     /// <blockquote>
     /// Normally you will use it without caring about these internals. <br/>
-    /// Please check @Articles.DynamicEntity
+    /// Please check @HowTo.DynamicCode.DynamicEntity
     /// </blockquote>
     /// </summary>
     [PublicApi]
@@ -53,10 +53,11 @@ namespace ToSic.Sxc.Data
         new object EntityTitle { get; }
 
         /// <summary>
-        /// Get a value of the entity. Usually you will NOT use this, because the default access
-        /// is usually content.FirstName - which will give you the same things as content.Get("FirstName").
-        /// The only reason to use this is when you dynamically assemble the field name in your code, like
-        /// when using App.Resources or similar use cases. 
+        /// Get a value of the entity. Usually you will prefer the quick access like
+        /// @content.FirstName - which will give you the same things as content.Get("FirstName").
+        /// There are two cases to use this:
+        /// - when you dynamically assemble the field name in your code, like when using App.Resources or similar use cases.
+        /// - to access a field which has a conflicting name with this object, like Get("Parents")
         /// </summary>
         /// <param name="name"></param>
         /// <returns>An object which can be either a string, number, boolean or List&lt;IDynamicEntity&gt;, depending on the field type. Will return null if the field was not found. </returns>
@@ -85,17 +86,33 @@ namespace ToSic.Sxc.Data
         /// <remarks>New in 10.06</remarks>
         new bool IsDemoItem { get; }
 
+        #region parents / children
+
         /// <summary>
         /// A dynamic list of entities which point to this item. Important for LINQ style querying or just
-        /// working with various lists. Note that for getting child items of this item you should just use the properties, like
-        /// content.Authors. 
+        /// working with various lists. Note that for getting child items of this item you
+        /// can just use the properties, like content.Authors. <br/>
         /// Please check the tutorials on 2sxc.org/dnn-tutorials/ for more info. 
         /// </summary>
         /// <param name="type">Optional type filter - would only return items of this type. </param>
         /// <param name="field">Optional field filter - would only return items that point to the current item in a specific field name.</param>
         /// <returns>A list of all items pointing here (filtered), converted to DynamicEntity for convenience.</returns>
-        /// <remarks>New in 9.42</remarks>
-        new List<IDynamicEntity> Parents(string type = null, string field = null);
+        /// <remarks>New in 9.42 - note also that the parameter-order is reversed to the Children()</remarks>
+        List<IDynamicEntity> Parents(string type = null, string field = null);
+
+        /// <summary>
+        /// A dynamic list of sub-items. Important for LINQ style querying or just
+        /// working with various lists. Note that for getting child items of this item you
+        /// can just use the properties, like content.Authors. <br/>
+        /// But using Children("Authors", typeName) gives you the ability to restrict to a type.  <br/>
+        /// Please check the tutorials on 2sxc.org/dnn-tutorials/ for more info. 
+        /// </summary>
+        /// <param name="type">Optional type filter - would only return items of this type. </param>
+        /// <param name="field">Optional field filter - would only return items that point to the current item in a specific field name.</param>
+        /// <returns>A list of all items pointing here (filtered), converted to DynamicEntity for convenience.</returns>
+        /// <remarks>New in 10.21.00 - note also that the parameter-order is reversed to the Parents()</remarks>
+        List<IDynamicEntity> Children(string field = null, string type = null);
+        #endregion 
 
         /// <summary>
         /// Contains presentation settings for an item - if they exist.